extract reference count analysis
